Story/route 401

+++ src/Workshop.Api/Bll/Models/CalculationLogModel.cs
@@ -1,7 +1,7 @@
 ﻿namespace Workshop.Api.Bll.Models;
 
 public record CalculationLogModel(
-    double Volume,
-    double Price,
-    double Weight,
-    double Distance);
+    decimal Volume,
+    decimal Price,
+    decimal Weight,
+    decimal Distance);
