+++ src/Workshop.Api/Bll/Services/Interfaces/IPriceCalculator.cs
@@ -4,7 +4,7 @@ namespace Workshop.Api.Bll.Services.Interfaces;
 
 public interface IPriceCalculator
 {
-    double CalculatePrice(IEnumerable<GoodModels> goods, double distance);
+    decimal CalculatePrice(IEnumerable<GoodModels> goods, decimal distance);
 
     IEnumerable<CalculationLogModel> QueryLog(int take);
