+++ src/Workshop.Api/Responses/V2/Report01Response.cs
@@ -1,8 +1,8 @@
 ﻿namespace Workshop.Api.Responses.V2;
 
 public record Report01Response(
-    double MaxWeight,
-    double MaxVolume,
-    double MaxDistanceForHeaviestGood,
-    double MaxDistanceForLargestGood,
-    double WavgPrice);
+    decimal MaxWeight,
+    decimal MaxVolume,
+    decimal MaxDistanceForHeaviestGood,
+    decimal MaxDistanceForLargestGood,
+    decimal WavgPrice);
